CMSSW_8_0_0

+++ MELA/test/testME_v2.c
@@ -115,7 +115,7 @@ void testME_FullMELA_PingWithFourMomenta(int flavor=0){
       GenLep3Id=13;
       GenLep4Id=-13;
     }
-    int idOrdered[4] ={ GenLep1Id, GenLep2Id, GenLep3Id, GenLep4Id };
+    int idOrdered[4] ={ static_cast<int>(GenLep1Id), static_cast<int>(GenLep2Id), static_cast<int>(GenLep3Id), static_cast<int>(GenLep4Id) };
     TLorentzVector pOrdered[4];
     pOrdered[0].SetXYZT(l1_array[ev][1], l1_array[ev][2], l1_array[ev][3], l1_array[ev][0]);
     pOrdered[1].SetXYZT(l2_array[ev][1], l2_array[ev][2], l2_array[ev][3], l2_array[ev][0]);
